+++ NEWS.md
@@ -1,4 +1,6 @@
 # FSA 0.7.1 ongoing
+* `ageBias()`: Modified.  Moved into a single file with `agePrecision()`.  No change in behavior.  Cleaned-up help file.
+* `agePrecision()`: Modified.  Moved into a single file with `ageBias()`.  No change in behavior.  Cleaned-up help file.
 * `hist.formula()`: Modified.  Fixed so that `ymax=` also sets the y-axis limit when only one histogram is made (it was previously ignored).
 * `iHndlFormula()`: Modified.  Added `na.action=NULL` to `model.frame()` so that NA values will not be omitted.
 * `psdAdd()`: Modified.  Fixed bug that never tested if all lengths were `NA`.  Required change to `iHndlFormula`.